Extract game logic from API
