+++ pages/api/games/[uuid].js
@@ -1,12 +1,6 @@
+import { applyMoveOnGame } from "../../../engine/game";
+import { getCurrentPlayer } from "../../../engine/player";
 import { getGameRepository } from "../../../models/games/gameRepository";
-import {
-  getCurrentPlayer,
-  WINNER_LINE_VALUE,
-  NO_PLAYER_VALUE,
-  FIRST_PLAYER_VALUE,
-  SECOND_PLAYER_VALUE,
-} from "../../../engine/player";
-import { getWinningPath } from "../../../engine/game";
 
 export default (req, res) => {
   const method = req.method;
@@ -18,7 +12,6 @@ export default (req, res) => {
       return patch(req, res);
   }
 };
-import _ from "lodash";
 
 function get(req, res) {
   const {
@@ -61,6 +54,7 @@ function patch(req, res) {
 }
 
 /**
+ * This function will update a game state, save it and return it into a JSON response.
  *
  * @param {integer} player
  * @param {integer} cellIndex
@@ -75,80 +69,16 @@ function playMove(player, cellIndex, uuid, res) {
       },
     })
     .then((game) => {
-      const grid = JSON.parse(game.grid);
-      console.log(grid);
+      const updatedGame = applyMoveOnGame(game, player, cellIndex);
 
-      if (game.secondPlayerNickname && player === getCurrentPlayer(grid)) {
-        const updatedGrid = grid.map((hexagon, index) =>
-          cellIndex === index ? player : hexagon
-        );
-
-        const winningPath = getWinningPath(updatedGrid, player);
-        if (winningPath) {
-          game = getWonGame(game, grid, player, winningPath);
-          console.log(game);
-        } else {
-          game.player = getNextPlayer(player);
-          game.grid = JSON.stringify(updatedGrid);
-        }
-
-        game.save().then(function (updatedGame) {
-          updatedGame.grid = JSON.parse(updatedGame.grid);
-
-          res.status(200).json(updatedGame);
-        });
-      } else {
-        res.status(400).json({ error: "NO_SECOND_PLAYER_OR_NOT_YOUR_TURN" });
-      }
+      updatedGame.save().then(function (updatedGame) {
+        updatedGame.grid = JSON.parse(updatedGame.grid);
+        res.status(200).json(updatedGame);
+      });
     })
     .catch((message) => {
       res.status(404).json({ error: `Game not found or ${message}` });
     });
-}
-
-/**
- * This function returns the Game object with an updated winning grid.
- * Also set the current player as "no player" and the winner as current player
- *
- * @param {Object} game
- * @param {Array} grid
- * @param {integer} player
- * @param {Array} winningPath
- */
-function getWonGame(game, grid, player, winningPath) {
-  game.grid = JSON.stringify(
-    grid.map(function (value, index) {
-      return hexagonIndexIsInPath(winningPath, index)
-        ? WINNER_LINE_VALUE
-        : value;
-    })
-  );
-
-  game.player = NO_PLAYER_VALUE;
-  game.winner = player;
-
-  return game;
-}
-
-/**
- * Check if the index of an hexagon is in the winning path.
- *
- * @param {[]} winningPath
- * @param {int} index
- */
-function hexagonIndexIsInPath(winningPath, index) {
-  return _.indexOf(winningPath, (index + 1).toString(10)) >= 0;
-}
-
-/**
- * Determines who will play next.
- *
- * @param {*} player
- */
-function getNextPlayer(player) {
-  return player === FIRST_PLAYER_VALUE
-    ? SECOND_PLAYER_VALUE
-    : FIRST_PLAYER_VALUE;
 }
 
 /**
